Create dedicated stack to handle the backfill data.
The backfilling data can cause a delay of a few days in the real-time data due to the backlog of Pub/Sub.
This change is to set up a dedicated Pub/Sub + ingester for data backfilling. This will ensure that the live data is not delayed by the historical data. Once the historical data is fully ingested, we can shut down the dedicated Pub/Sub + ingester

Change-Id: Ia613024cb6074596c2896efc12f56f63b7c33112
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/712036
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Alex Sun <sunpeng@google.com>