[jsfiddle] Create jsfiddle_container-base container
Create a new jsfiddle_container-base Docker image. This image
includes all but the Skia bits. The Skia artifacts (CanvasKit and
PathKit) will be added in a later step to create the final image.
This is a necessary step to building these applications in Louhi.

Bug: b/295534733
Change-Id: Icf96fe599a859f808ff5bff3db73c67efaa9b41d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/740696
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Chris Mumford <cmumford@google.com>