[perf] Add settling time to regression detection.
SettlingTime is the amount of time to wait before including data from
a commit in Anomaly detection.
For example, because of machine contention or retries for some tests,
the results may arrive out of order causing Anomalies to be
mis-attributed, or attributed to a series of different CLs as
new data arrives.

Also increase test coverage of dfiter.go to 86%.

Finally set a settling time for the android-x instance to 5 hours.

Bug: b/40045434
Change-Id: I762564a7f5dcc141d0344901b0a83b7fb64c48cc
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/745156
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Leandro Lovisolo <lovisolo@google.com>