[NPM-audit-mirror] Integrate with Buganizer and remove support for Monorail
NPM audit mirror supported filing Monorail bugs in any Monorail instance and component, we are going to replace that with filing Buganizer bugs only in the "Skia Infrastructure > NPM-audit-mirror" component. Owners can then move the bug into their own Buganizer component if needed.

As part of this work, the markdown in the filed bugs was improved. Examples of bugs filed-
* Audit issue:  https://g-issues.skia.org/issues/297927299
* Downloaded packages issue:  https://g-issues.skia.org/issues/297952058

Tested with unit tests and end-to-end tests in staging.

Bug: b/287255102
Change-Id: If553eb36c85622b3312103aca90eb31ae382dabb
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/747797
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Ravi Mistry <rmistry@google.com>