[gold] digest-details-sk: Don't crash when digest is outside of current tile.
When querying the details of a digest outside of the current tile, the /json/v2/details RPC returns a SearchResult with an empty paramset. This can be attributed to the getTracesForGroupingAndDigest function here: https://skia.googlesource.com/buildbot/+/76d21d4985f87074b213f6234ffe8348d81daaac/golden/go/search/search.go#3674.

From the paramset, digest-details-sk constructs a grouping with getGrouping() that is subsequently passed to dots-legend-sk, which needs a grouping to render links to the diff and details page. The reason why digest-details-sk crashes is that getGrouping() throws an exception if it cannot find the param keys required for a grouping, which is the case when the paramset is empty.

To work around this problem, digest-details-sk has checks in various places to conditionally render some elements based on whether getGrouping() was successful (search for "maybeGrouping" in digest-details-sk.ts). However, we don't perform that check before rendering dots-legend-sk. This CL adds a try/catch around getGrouping() like we do in other places, and only renders dots-legend-sk if getGrouping() was successful.

Bug: b/293619865
Change-Id: Id9d20353acc2995097256a69365d54980e2a2350
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/756039
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>
Auto-Submit: Leandro Lovisolo <lovisolo@google.com>