[datahopper] Add Skia repo to Louhi ingester, add more error details
We're now building images in Louhi from the Skia repo, and the Louhi
ingester in datahopper needs to retrieve commit information. Add the
flow name to the error if we fail to retrieve commit details so that we
have a trail of breadcrumbs to follow if this fails again.

Change-Id: I6782fe3e5a58003f89aa9ac6a52549f3a29f030a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/767540
Reviewed-by: Ravi Mistry <rmistry@google.com>
Commit-Queue: Eric Boren <borenet@google.com>