[task scheduler] Split initial try job DB insert from sync
The current flow is to query Buildbucket V1 for requested builds, then
perform potentially-expensive syncing and caching before leasing the
build and inserting into the database. This CL changes to query
Buildbucket V1 for requested builds, insert them into the DB with
JOB_STATUS_REQUESTED, and then perform the syncing and caching
subsequently in another thread. This prepares us for the Buildbucket V2
migration described in go/skia-bb-taskbackend, which  the insertion into
the DB will be moved to an HTTP frontend.

Bug: b/288158829
Change-Id: Iff8b9c89154b5eab6c54b69f6c53f1dc209c7d5e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/780099
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Eric Boren <borenet@google.com>