[task scheduler] Add more logging to diagnose ConcurrentUpdate errors
This is the old flow and it's disappearing soon, but the errors are
concerning me despite the fact that nothing appears outwardly wrong. My
best guess is that we're seeing the same build IDs appear repeatedly in
Peek(), which causes us to create duplicate jobs for the same build
which then fail when we try to Start them. I don't know why this is
producing the ConcurrentUpdate errors though.

Change-Id: I99ed405d0ee11ee28d07241f0a94b1257d6ef5aa
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/790557
Commit-Queue: Eric Boren <borenet@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>