[cabe] use NormalizeResult instead of LogTransform if data has zeros
This is a fix for handling data that currently causes
errors in prod, which in turn causes sandwich verification to fail safe
and proceed with full bisections of regressions that probably don need
to be bisected in the first place.

The heuristic is simple: if either the treatment or control measurements contain any zeroes, use the NormalizeResult transform. Otherwise, continue to use the LogTransform as usual.

To demo, run this command with and without this CL:
```
bazelisk run //cabe/go/cmd/cabe -- analyze  --pinpoint-job 11a4b863ce0000
```

*Without this CL*: it will fail with an error message like:
"Error: problem reported by cabe_stats.BerfWilcoxonSignedRankedTest: x[i] and y[i] must be greater than 0 to use log transform, but got x[i]: 1.5 and y[i]: 0"

*With this CL*: it will complete the analysis with no errors and the output will indicate only one difference: a minor improvement in a metric that wasn't event part of the sandwich verification request. So cabe would have responded to the sandwich regression verification step in https://bugs.chromium.org/p/chromium/issues/detail?id=1515888 with a "NoRepro" rather than an error that forced it to proceed with a full bisection.



Change-Id: I3b04b63d04d6e68978db625977801dd6c059b38d
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/798502
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Alex Sun <sunpeng@google.com>
Commit-Queue: Sean McCullough <seanmccullough@google.com>