Update Slot Manager to persist slot changes to JSON
This updates the Slot Manager to use the same pattern as text editor of pushing changes back to the main component as they happen. This leverages existing code in the main component to detect changes as well as re-render the player. This means slot manager no longer needs to apply changes directly to the player. Because the player does a full re-render, this will also fix the bug where slot changes weren't applying to static/paused animations.

Bug: b/313642637
Bug: b/313641683
Change-Id: I6a217417a3ed080f2fc5b82ce65d939898af6531
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/801917
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Brandon Wilson <bwils@google.com>