Use service proto request definition for pinpoint.go
This change uses the proto service Pinpoint request in favor of
the existing request definition in pinpoint.go in preparation for
transitioning things for Temporal workflows.

This adds aggregation method to the request.

Note that this does not port over the Response struct yet because
there are ties to the various structs that would also need to
be moved over.

Bug: b/324487856
Change-Id: I689c980812c0944b848f4b711ffe22e9ae899f38
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/810193
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>