Improve the PerflightQuery to reduce latecy for the /count api.
Currently whenever a param-value is clicked in the query dialog, a request goes to the /_/count api that returns the number of traces for the current selection as well as available options in the other params based on currently selected options. This call takes too long for chrome instances(5~10 seconds for the public instance, 50~60 seconds for the internal instance) due to the sequential database query.
This CL will change the sequential query to parallel query to speed up the /_/count API.

Bug: https://b.corp.google.com/issues/312227681

Change-Id: Ic40f789f1bbf525d1bbe52d8d6305a26da751d92
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/812720
Commit-Queue: Alex Sun <sunpeng@google.com>
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>