Add FindMidCommit and remove CombinedCommit and CommitRange.
FindMidCommit is a simpler API to only return a middle Commit from the
given two commits. CommitRange is not used and can be constructed on
its own, so it is removed. CombinedCommit is only useful to build
Chrome, the Commit itself can identify where the middle point is.

Bug: b/26352320
Change-Id: I4d1c8e7bee7049a2cef69c799fd284731ab9f0a4
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/817356
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>