Add BisectWorkflow skeleton.
This is the skeleton version where we only do 50% of things we should do. This version only finds the first culprit and it doesn't compile the build with modified deps.

This is mainly for the skeleton so other activities and workflows can
be individually worked on.

Change-Id: I9fb6dc7f480317b0176ccc4efbcf5cce6ff0b63f
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/813872
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>