Add pinpoint prefix back to proto.
chi doesn't rewrite the request path and the json handler will see the
original url path. The handler sees the prefix and need to be able to
recognize it so we need to add the prefix in the proto definition.

Bug: b/322047067
Change-Id: I422769a521cc936ec1a010d96f3fc75b147355ee
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/822608
Auto-Submit: Hao Wu <haowoo@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>