Properly load web fonts when uploading a new file.
Some of the loading paths were dependent on the hash being set but were being run before that was available due to eager loading and erroring. Fix those to be more resilient if the hash is not present.

This bug was likely made more apparent due to changes in how Skia was (no longer) doing font fallbacks. This also updates the pinned version of Skia for the local dev build.

Bug: b/329118857
Change-Id: Ie891224d42801c02677e3a123d605c3789c479c0
No-Presubmit: true
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/824828
Reviewed-by: Kevin Lubick <kjlubick@google.com>