[bisection] Implement FindMidCombinedCommit
This function calculates the midpoint between two combined commits.

When ModifiedDeps are not specified, the function searches for the
midpoint in the range of Chromium commits. If the start/end Chromium
commits are adjacent, a DEPS roll is assumed and the DEPS files at
start and end commits are compared to find the rolled DEP, and the
logic will fetch the midpoint from that rolled DEP. If the rolled DEP
is also adjacent, the method stops and returns nil, saying there's
nothing further to dig through.

If ModifiedDeps are defined, the function will first find the modified
depdendency that has a different git hash between start and end. That
dependency will be used to find the midpoint. For example, if we have
start = chromium@1, v8@2 and end = chromium@1, v8@5, it will find the
midpoint using v8, between 2 and 5. If the logic reaches a point where
the V8 commits are adjancent, again, a DEPS roll is assumed and it
searches in DEPS files.

There comes a point when running comparisons where one combined commit
may be missing modified deps. For example, if chromium commits are
adjancent and DEPS files are parsed, we may want to find the midpoint
between chromium@1 vs chromium@1, v8@3. If chromium@1's DEPS file has
V8@1, the combined commit needs that information to perform an apples
to apples comparison for v8@1 adn v8@3. Thus, fillModifiedDeps() is
implemented with this change to accomplish that.

This change also updates LogLinear -> LogFirstParent, which is what we
were actually looking for and is what catapult's pinpoint is also
doing.

Bug: b/326352320
Change-Id: If811393e46d469b765f3b5bee4b70af3cd53bd58
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/822607
Reviewed-by: Hao Wu <haowoo@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>