[bisection] return startcommit when deps are equal
Update findMidCommitInDEPS to return null when the two DEPS files are
equal. FindMidCombinedCommit is updated to return startCommit when
null is hit, and the workflow is updated to check the Key() of lower
and the mid to see if they are equal to terminate instead of null.

Bug: b/326352320
Change-Id: I8b944be7091f63ce28fce1404be8fce7186354eb
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/826204
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>