Reland "[bisection] return startcommit when deps are equal"
This is a reland of commit 8d4395cdf735c67056dd1d386bbc54a9034154ea.

Changes:

Responses are consolidated to return startCommit as indicator to
workflow for termination.

findMidCommit(), the common coordinator for the midpoint search,
masks that findMidCommitInDEPS() can return:
1. nil, if DEPS files are equal and there is no valid midpoint
2. startCommit, when the two commits in DEPS files are also adjacent.

This adds a note that findMidCommitInDEPS() should also assume
a DEPS roll for #2, but that is not implemented as part of this change.

Thus, FindMidCombinedCommit() also returns nil as a termination clause
and the workflow only needs to perform a nil check instead of
verifying that it's the same as start commit for now.

Original change's description:
> [bisection] return startcommit when deps are equal
>
> Update findMidCommitInDEPS to return null when the two DEPS files are
> equal. FindMidCombinedCommit is updated to return startCommit when
> null is hit, and the workflow is updated to check the Key() of lower
> and the mid to see if they are equal to terminate instead of null.
>
> Bug: b/326352320
> Change-Id: I8b944be7091f63ce28fce1404be8fce7186354eb
> Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/826204
> Commit-Queue: Jeff Yoon <jeffyoon@google.com>
> Reviewed-by: Hao Wu <haowoo@google.com>

Bug: b/326352320
Change-Id: I1f6260e4bcd3a85630ee2882e7aa9cfdd8e96e7e
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/826577
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>