[pinpoint] Return commit in GetAllValues to debug in temporal UI
In the temporal UI, there is no easy way to infer the two commits prior
to the CompareActivity if they are considered the same. Given that
commit pairs are running out of order it would be nice to make the UI
workflow easier to follow.

Bug: b/326352319
Change-Id: I07cc220b518455d110c3399c1d78da07c610b2bd
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/826677
Reviewed-by: Hao Wu <haowoo@google.com>
Auto-Submit: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>