[bisection] don't reverse order for calculating midpoint
The gitiles response returns in reverse chronological order, so the
code was implemented to calculate midpoint in chronological order
by reversing the giitles response.

To retain as close behavior to Catapult Pinpoint, this is reverted,
so the next candidate is calculated from the list of reverse
chronological order.

Change-Id: I355849d938738e30384cbaee8d6e1e6c3d0cb638
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/826943
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>