Refactor Bisection workflow to use BisectRun to track progress.
The separate map[hash]*CommitRun is replaced by a new struct BisectRun.
BisectRun directly tracks the commit and its CommitRun result as the
lookup in the table is not necessary when the actual commits are already
captured in the CommitRangeTracker. Each compare between commits are
also independent so no lookup is needed.

This CL also moves this new struct into a new file and a separate
unit test file. The newly added test cases should cover the existing
cases.

This CL is also a preparation for running bisection concurrently.

Bug: b/326352379
Change-Id: I4bc9556f37bda6e0cc943b33947be75b02d215a0
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/828956
Commit-Queue: Hao Wu <haowoo@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>