Add bisectRunTracker to be able to serialize CommitRangeTracker.
bisectRunTracker maps indexes to the actual BisectRun such that the
indices can be used for CommitRangeTracker and be able to serialize.
Note workflow.Future is a runtime feature that can't be serialized.

The underlying BisectRun needs one lookup and it will be shared
between each comparisons. This prepares the future work to run
bisection comparisons in parellel.

Bug: b/326352379
Change-Id: I89899099f2ee4acc701b38f24e85e12bde8341d1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/830617
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>