Move build params into Build.
The params including the commit is added to the workflows.Build so
each build is now associated with the original params that is used
to make builds. CombinedCommit in SingleCommitRunnerParams can be
replaced with the Build but will be done in a separate CL.

The workflows.Build then can identify all the contextual information
that's used to make it.

This CL also changes to pass by value so it has better intention in
the function signature and less error-prone to nil pointers.

Change-Id: I7fdd39cbc2b1f1bbcedc090a29c9ad39a1faba3b
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/832146
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Hao Wu <haowoo@google.com>