[bisect] Don't assume DEPS file exists
The existing logic assumed that DEPS file would always be present.
The logic is updated to terminate midpoint calculation when
DEPS files are not present on either side of the commit.

The client to midpointHandler specifies With2xxOnly, meaning that
non 200 based errors will error. This has been removed with this
change.

However, there still exists the check in gitiles.go get() function,
which enforces a 200 response code. We check for that error
specifically through a string match as a workaround to check for
404.

Bug: b/331408673
Change-Id: I7000c87678aaef28fb47b2a2f244a4405a0c9168
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/832917
Reviewed-by: Hao Wu <haowoo@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>