[pinpoint] introduce improvement direction to compare
When bisection compares two commits, the statistical tests are not
concerned with the directionality of the change. So it is possible
for improvements to be identified as regressions. Add improvement
direction to compare to filter out improvements.

This change will not affect the current bisect workflow. Improvement
direction needs to be passed in before this change can take effect.

Bug: b/321305126
Change-Id: I3a26ba272dba365fc5e24f30d0f845da7779a012
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/833956
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Reviewed-by: Jeff Yoon <jeffyoon@google.com>