Define CombinedCommit, Commit and ModifiedDeps in proto
The current bisection response is []string for the hashes, but we
want to capture the entire combined commit object as culprits.
The bisection response is defined in proto, so this change moves the
object definitions from midpoint to proto and are generated.

See note in bisect.go for type modifications while using combined
commit from midpoint, and storing the result via pb.combined commit
as a result of this change.

Change-Id: I4f11a5dd02588322f48f11d22e80887774d27163
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/833571
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>