Clean up fiddle's tests and use of k8s namespace
While looking at fiddle before moving it to skia-infra-public, I realized
the k8s namespace it looks at is hardcoded to use "default".

If we ever want to move fiddle and fiddler to the skia-infra-public
project, we'll want to put it in a namespace (e.g. fiddle, which
I've already made). Thus, we need to configure that option.

While I was plumbing things through, I made the code and tests
conform to a more modern style.

Change-Id: Ie3469763bdc4dc1f87483db2590ae5885021b71c
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/835990
Reviewed-by: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Kevin Lubick <kjlubick@google.com>