[bisect] IssueTracker Client and integration into Bisect workflow
go/backends/issuetracker contains the logic and handling for
IssueTracker. Currently, it supports writing a comment to IssueTracker
for when culprits are found. The template for this comment is
outlined in culprit_detected.tmpl, and is instantiated when the
IssueTracker object is generated.

The bisect workflow is updated to use the IssueTracker client and
make updates on the comments when the BugId is provided. It currently
only logs failures to ensure that the workflow fails. Newer tickets
created should have pinpoint-worker SA added as a collaborator,
but older tickets may not have it, meaning that we're unable to make
edits to the ticket.

Bug: chromium:322047308
Change-Id: I26250e4e865977bf0e78faaf67f324fe8b4793e9
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/833563
Reviewed-by: Kevin Lubick <kjlubick@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>
Reviewed-by: Leina Sun <sunxiaodi@google.com>