[pinpoint] quality of life changes to GetAllValuesLocalActivity and CompareActivity
When viewing a temporal workflow from the UI, every compare generates
6 lines of activities. This CL condenses GetAllValuesLocalActivity and
GetAllErrorValuesLocalActivity into a single local activity. This CL
also combines functional and performance comparisons into a single
activity. This way we minimize noise and only generate 2 lines per
comparison.

This CL also appends the commit information to the CompareActivity so
that it is easier to identify the two commits that are being compared

Bug: b/327033086
Change-Id: I2e8329d00fd270b1eecbd4b0848f411d09448a06
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/833830
Commit-Queue: Leina Sun <sunxiaodi@google.com>
Reviewed-by: Hao Wu <haowoo@google.com>