[perf] Do not allow unclicking all kvp's on explore page.
From the linked bug, unchecking all params of a list makes the plot go away, so we need to detect when we are down to just one
checked value for a key in the paramset and at that point we
need to disable that final checkbox so it can't be unchecked.

This requires adding an `unchecked: Map<string, Set<string>>` to
the paramset-sk to track which values are unchecked. Note we
used 'unchecked' because all checkboxes start off as checked.

Bug: b/331701145
Change-Id: I130120152b61c19d52dfb48875712f4e82613f28
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/835186
Reviewed-by: Ashwin Verleker <ashwinpv@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>