[perf] Separate behavior for checkbox and label on paramset-sk.
That is, the checkbox-sk will produce the an event for clicking
on the checkbox, but the label will produce it's own event, which
is the one for clicking on 'values' in a paramset-sk.

This allows turning on both `clickable_values` and
`checkbox_values` on the explore page, which means you can
hide traces by unchecking the checkbox, and highlight by clicking
on the value, which then allows using the "Remove Highlighted"
and "Highlighted Only" buttons.

Change-Id: Ib9ae8b816d020545fc89be5e48fef17fe612f37a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/838144
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>