[perf] Fix zooming when the mouse leaves the plot window.
When the mouse left the plot this would trigger a mouseleave
event, which acted like a mouseup event and would
complete the zoom. But because there was a mousedown
that had started the zoom a 'click' event would also
be triggered which would highlight a trace and that in
turn would update the URL.

The fix here is to call ` e.preventDefault()` for
all the mouse events, which stops the click from
being generated.

Bug: b/333405264
Change-Id: I434ca8a520d7f740a41a46d2de53c52c187e593a
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/838905
Commit-Queue: Joe Gregorio <jcgregorio@google.com>
Auto-Submit: Joe Gregorio <jcgregorio@google.com>
Commit-Queue: Ravi Mistry <rmistry@google.com>
Reviewed-by: Ravi Mistry <rmistry@google.com>