[pinpoint] Create a new CatapultBisect workflow
The CatapultBisect workflow is in its own package, catapult,
that's intended to contain all backwards compatibility-related
logic in it. It utlizes internal.BisectWorkflow and the Catapult-based
workflow is planned to be extended to:
1. translate the raw data into the Legacy response object
2. call skia-bridge to store it in Catapult's Datastore.

Once the backwards compatibility is no longer needed, this package
should be removed and internal.BisectWorkflow should be used.

Bug: b/335543111
Change-Id: I9cccf67a43e34d0638859a543e6b4c5399d095d1
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/842433
Reviewed-by: Hao Wu <haowoo@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>