[pinpont] expand bisectexecution to return raw data
We need raw data surfaced in internal.BisectExecution to process
it into the Legacy repsonse format.

Creating a new struct with the proto BisectExecution embedded fails
to store the additional fields needed. To keep impact minimal to the
existing solution, we create a new struct similar to the proto
definition with the additional fields needed for backwards
compatibility. The compare workflow is updated to store all values
as part of its results.

Bug: b/335543410
Change-Id: I8f877ee03705a22f9222ed6ac27a62b1db7ba7a5
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/842434
Reviewed-by: Leina Sun <sunxiaodi@google.com>