[pinpoint] convert botID dimensions from slice to map
In pairwise workflow, we need to provide the bot ID in the dimensions
field of the swarming request when we run a benchmark. Previously the
dimensions were passed in as a slice, but the slice only contains
one element so a map is sufficient.

Bug: b/321304277
Change-Id: Ia3016c7ca988ccdf88e296f9cd31c0b84093aec6
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/841447
Reviewed-by: Jeff Yoon <jeffyoon@google.com>
Commit-Queue: Leina Sun <sunxiaodi@google.com>