[pinpoint] process run data into activities per commit
In the Legacy response there is one state object generated
per combined commit. The logic parses through run data and
sorts it into attempts, where each attempt consists of 3
executions, one for each quest.

A FindBotActivity is introduced to call swarming in order
to get the bot id of the associated task.

Bug: b/335543316
Change-Id: Icb2466e1f6a98e8060c300492ebd7fa60596d8fa
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/842941
Reviewed-by: Leina Sun <sunxiaodi@google.com>