[pinpoint] Update state objects with comparisons
Compare the performance results of each state object and update the
result values for each. For the first index, prev is unset and next
will be the result of comparing its values against the next index.
For the next iteration, prev = the comparsion result from before, and
next is recalculated. For the last index, only prev is set.

Bug: b/335543541
Change-Id: I2d66722b05d655212e5671a397659411bee01930
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/843996
Reviewed-by: Leina Sun <sunxiaodi@google.com>
Commit-Queue: Jeff Yoon <jeffyoon@google.com>