[fiddle] Fix rendering of images for the default fiddle.
Even if the fiddle was not an animation, but if it had a non-zero
animation duration, then it would try to render it as an animation,
but that would produce no CPU or CPU outputs, only ANIM_CPU and
ANIM_GPU outputs, so the fiddle would look like it failed to
draw anything.

Fix is to change the defaults for animation duration to 0, and
then to also only check for animations based on the Options.Animation
property, and not on Options.Duration.

Change-Id: I7c4be037ceeb5682f8d4a6562982cacf941c59cc
Reviewed-on: https://skia-review.googlesource.com/c/buildbot/+/845987
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Commit-Queue: Joe Gregorio <jcgregorio@google.com>